BrevkodePar-klassa har nær relasjon til EtterlatteBrevKode-enumen, så synes dei bør liggje i samme pakke
